fix(1.1-01): resolve all pre-existing clippy warnings across workspace
- vel-core: replace manual Default impls with #[derive(Default)] for
  PrivacyClass, SyncClass, ArtifactStorageKind (lib.rs) and AgentPriority (run.rs)
- vel-storage: add #[allow(clippy::too_many_arguments)] to 5 repository functions
  with legitimate many-param signatures (db.rs, commitments_repo.rs, nudges_repo.rs)
- veld: fix lifetime elision in middleware/mod.rs expected_token_for_class
- veld: mechanical fixes (consecutive replace, useless conversion, unnecessary cast,
  unnecessary closure, redundant closure) across adapters/health.rs, chat/messages.rs,
  inference.rs, nudge_engine.rs, suggestions.rs
- veld: add Default impl for WorkerRuntimeState in state.rs
- veld: add #[allow(clippy::too_many_arguments)] to 6 service functions
- veld: add #[allow(clippy::if_same_then_else)] to evaluate() in nudge_engine.rs
- veld: add #[allow(clippy::enum_variant_names)] to CommandExecutionPayloadKind
- vel-cli: fix empty format strings (inline literals), redundant closure,
  boolean simplification, consecutive replace, variant name suppression
- All 289 workspace tests pass, zero clippy warnings

+++ crates/veld/src/services/suggestions.rs
@@ -227,6 +227,7 @@ where
         .collect()
 }
 
+#[allow(clippy::too_many_arguments)]
 fn build_candidate(
     suggestion_type: SuggestionType,
     priority: i64,
@@ -527,7 +528,7 @@ fn rank_candidates(
             let age_days = (OffsetDateTime::now_utc().unix_timestamp() - newest_created_at)
                 .div_euclid(86_400)
                 .max(0);
-            (15 - age_days.min(15)) as i64
+            15 - age_days.min(15)
         } else {
             0
         };
